requirements.txt
streamlit==1.41.1
plotly==5.24.1
pandas==2.2.3
numpy==2.2.1
anthropic>=0.40.0
httpx>=0.24.0,<0.28.0
scipy==1.14.1
scikit-learn==1.6.1
python-dateutil==2.9.0
python-dotenv==1.0.0
reportlab==4.2.5
matplotlib==3.10.0
